% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2excel-internal.R
\name{html2excel-internal}
\alias{html2excel-internal}
\alias{is_url}
\alias{is_directory}
\alias{urls_files_or_directory}
\alias{distinguish_repeated_filenames}
\alias{change_xlsx_directory}
\title{Internal html2excel functions}
\usage{
is_url(x)

is_directory(x)

urls_files_or_directory(x)

distinguish_repeated_filenames(vec)

change_xlsx_directory(filename, dir = "output")
}
\description{
Internal html2excel functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
