% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2excel-package.R
\docType{package}
\name{html2excel-package}
\alias{html2excel-package}
\title{html2excel: Convert HTML Tables to Excel Files}
\description{
Reads tables from HTML web pages or local documents. The
tables are returned as a list of tibbles and may be written to Excel files.
}
\details{
The main function is \code{\link{html2excel}}.
See also \code{\link{summary.html2excel}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulnorthrop.github.io/html2excel/}
  \item \url{https://github.com/paulnorthrop/html2excel}
  \item Report bugs at \url{https://github.com/paulnorthrop/html2excel/issues}
}

}
\author{
\strong{Maintainer}: Paul Northrop \email{p.northrop@ucl.ac.uk} [copyright holder]

}
