% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2excel.R
\name{html2excel}
\alias{html2excel}
\title{Convert HTML Tables to Excel Files}
\usage{
html2excel(
  html,
  ext = "*.html",
  write = TRUE,
  dir,
  sheets,
  read_args = list(),
  html_args = list(),
  write_args = list()
)
}
\arguments{
\item{html}{A character vector containing one of the following:
\itemize{
\item paths to HTML files,
\item a path to a directory containing HTML files,
\item URLs, that is, strings beginning with \code{www.}, \verb{http:} or \verb{https:}.
}}

\item{ext}{A character scalar. If \code{html} is a path to a directory, then
\code{ext} determines the extensions of files that should be read from \code{dir}.
The default is only to read files with an \code{.html} extension.
\code{utils::glob2rx(ext)} is used to create the \code{pattern} argument to
\code{\link{list.files}}. For example, if \code{ext = "*.htm*"} then all files with
extensions that start with \code{htm} are read.}

\item{write}{A logical scalar. Should the tibbles be written to Excel
\code{.xlsx} files using \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx}}?}

\item{dir}{A path to a directory. If \code{write = TRUE} then \code{dir} determines
where the Excel files are written.
\itemize{
\item If \code{html} contains URLs, then \code{dir} is an absolute path and must be
supplied. To write Excel files to the current working directory use
\code{dir = "."}.
\item If \code{html} contains paths to HTML files or a directory containing HTML
files, then \code{dir} is relative to the directory in which an HTML file is
located. If \code{dir} is not supplied then Excel files are written to the
respective directories supplied in \code{html}.
}

If \code{dir} does not exist then it is created.}

\item{sheets}{A numeric vector or list of numeric vectors. Component \code{i}
of the vector or list gives the number(s) of the tables to be included as
sheets in the \code{i}th output Excel file. This argument is recycled to the
number of output Excel files. If \code{sheets} is not supplied then all sheets
are included. \code{\link{summary.html2excel}} provides dimensions of the tibbles
read.}

\item{read_args}{A list of arguments for \code{\link[rvest:read_html]{rvest::read_html}}.}

\item{html_args}{A list of arguments for \code{\link[rvest:html_table]{rvest::html_table}}.}

\item{write_args}{A list of arguments for \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx}}, but not
\code{file}, as this is determined from \code{html} and \code{dir}.}
}
\value{
An object of class \code{c("html2excel", "list")}. A list of (lists of)
tibbles created from objects returned from \code{\link[rvest:html_table]{rvest::html_table}}.
The names of the list are the input HTML filenames. Each list component
is named \code{sheet1}, \code{sheet2} etc. with each sheet containing a separate
table found in the input HTML file. The numbering of these sheets is the
same as the output \code{.xlsx} files, not the order of the tables in the
input HTML files.
If \code{write = TRUE} then a character vector containing the file paths to the
\code{.xlsx} files created is added as an attributed named \code{files}.
}
\description{
Reads tables from HTML documents using \code{\link[rvest:read_html]{rvest::read_html}} and
\code{\link[rvest:html_table]{rvest::html_table}}. The tables are returned as tibbles and may be written
to Excel \code{.xlsx} files using \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx}}.
}
\details{
HTML files are
\itemize{
\item read using \code{\link[rvest:read_html]{rvest::read_html}},
\item converted to tables (tibbles) using \code{\link[rvest:html_table]{rvest::html_table}},
\item written to Excel documents using \code{\link[openxlsx:write.xlsx]{openxlsx::write.xlsx}}.
}

If \code{\link[rvest:read_html]{rvest::read_html}} fails to connect to a URL in \code{html} then a message
is printed and \code{html} is returned.

The output filename for an input HTML file \code{file.html} is usually
\code{file.xlsx}. Exceptions are when a combination of \code{html} and \code{ext} means
that duplicate filenames would be produced, for example, if files
\code{file.html} and \code{file.mhtml} are read. Then the output filenames are
distinguished by \code{file_a.xlsx} and \code{file_b.xlsx}, and similarly if there are
more than two identical filenames.

The initial motivation for creating the \code{html2excel} package was to convert
to Excel format HTML files that has mistakenly been given a \code{.xlsx} file
extension. If such files are supplied by \code{html} then each output Excel
file will be written to a directory \code{output} created in the directory of the
respective input file.
}
\examples{
### HTML files

## .html and .mhtml versions of the URL example below
html <- system.file("extdata", "tables.html", package = "html2excel")
mhtml <- system.file("extdata", "tables.mhtml", package = "html2excel")

## .html
# Table 5 only
t1 <- html2excel(html, sheets = 5, write = FALSE)
# The table
t1[[1]]
# The dimensions of the table
summary(t1)

## .mhtml
# Pass UTF-8 encoding to avoid error "Unsupported encoding: 3DUTF-8"
t2 <- html2excel(mhtml, read_args = list(encoding = "UTF-8"), write = FALSE)
# The same table as above
t2[[1]][5]
# The dimensions of all tables
summary(t2)

### A directory

## Contains the .html and .mhtml files above
directory <- system.file("extdata", package = "html2excel")
# Change the ext argument to include the .mhtml file
# Extract tables 3 and 5 from .html and table 5 from .mhtml
x <- html2excel(directory, ext = "*.*html", sheets = list(c(3, 5), 5),
                read_args = list(encoding = "UTF-8"), write = FALSE)

### A URL
url <- "https://afd.calpoly.edu/web/sample-tables"
x <- html2excel(url, write = FALSE)
}
\seealso{
\code{\link{summary.html2excel}}
}
