% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_method.R
\name{summary.html2excel}
\alias{summary.html2excel}
\alias{print.summary.html2excel}
\title{Summarising \code{html2excel} objects}
\usage{
\method{summary}{html2excel}(object, ...)

\method{print}{summary.html2excel}(x, ...)
}
\arguments{
\item{object}{An object inheriting from class \code{"html2excel"} returned from
\code{\link{html2excel}}.}

\item{...}{Further arguments to be passed to \code{\link{print.default}}.}

\item{x}{An object returned by \code{summary.html2excel}.}
}
\value{
A list containing the dimensions (number of rows and number of
columns) of each tibble.
}
\description{
\code{summary} method for class \code{html2excel}.
}
\details{
The default \code{print} method for a list of tibbles prints the
dimensions of the tibbles and up to 10 rows and 6 columns of each tibble.
\code{summary.html2excel} prints a shorter summary that contains only the
dimensions of the tibbles. These \code{print} and/or \code{summary} methods may be
helpful in deciding which sheets are required for which excel file, that
is, how to set the argument \code{sheets} to \code{\link{html2excel}}.
}
\section{Examples}{
 See \code{\link{html2excel}}.
}

\seealso{
\code{\link{html2excel}}.
}
